requirements.txt
yfinance
openai
newsapi-python
pydantic
finnhub-python
SQLAlchemy